# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dstdelivery",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/add/toCart": {
            "post": {
                "tags": [
                    "Cart Section"
                ],
                "summary": "Add To Cart",
                "description": "",
                "operationId": "b4d9ac5db4bb0b0971d508b77b9542ca",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Please write product ID",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Please write user ID",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "totalQty",
                        "in": "query",
                        "description": "Please write totalQty",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/get/cart/products": {
            "get": {
                "tags": [
                    "Cart Section"
                ],
                "summary": "Get Cart Products",
                "description": "",
                "operationId": "5109a3eebfcb7725835c2243d3c3910f",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Please write user ID",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/delete/cart/products": {
            "post": {
                "tags": [
                    "Cart Section"
                ],
                "summary": "Delete Cart Products",
                "description": "",
                "operationId": "13a351452c12d0f6ceebfd54d462cfbf",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Please write product ID",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/filter/cart": {
            "get": {
                "tags": [
                    "Cart Section"
                ],
                "summary": "Request that search via product name",
                "description": "",
                "operationId": "868a499852e6d2c211c9a220ce31ee8d",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Please write cart name",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}